UPSTREAM: devtmpfs: don't mix {ramfs,shmem}_fill_super() with mount_single()
Create an internal-only type matching the current devtmpfs, never
register it and have one kernel-internal mount done.  That thing
gets mounted only once, so it is free to use mount_nodev().

The "public" devtmpfs (the one we do register, and only after
the internal mount of the real thing is done) simply gets and
returns an extra reference to the internal superblock.

Change-Id: I784d71aa1505396319989cd402019f7cab97ae22
Signed-off-by: Al Viro <viro@zeniv.linux.org.uk>
Signed-off-by: Aditya Eka Pranata <aepranata@gmail.com>